Fix withdraw Solana event types

+++ everscale/contracts/bridge/interfaces/IProxyTokenTransfer.sol
@@ -24,6 +24,6 @@ interface IProxyTokenTransfer {
         TvmCell code
     ) external;
 
-    event WithdrawSolana(address addr, uint64 tokens, uint256 solana_addr);
+    event WithdrawSolana(address addr, uint128 tokens, uint256 solana_addr);
     event WithdrawEthereum(int8 wid, uint256 addr, uint128 tokens, uint160 eth_addr, uint32 chainId);
 }
